Move quit to ESC+ALT

+++ L2Alone/L2Alone/L2QuitKeyHandler.h
@@ -8,7 +8,7 @@
 class L2QuitKeyHandler : public L2KeyboardEventHandler {
 public:
 
-	L2QuitKeyHandler(DWORD l2ProcessId, vector<L2AccountHotKey> *accountHotKeys) {
+	L2QuitKeyHandler(DWORD l2ProcessId, vector<L2AccountHotKey>* accountHotKeys) {
 		this->l2ProcessId = l2ProcessId;
 		this->accountHotKeys = accountHotKeys;
 	}
@@ -35,30 +35,18 @@ public:
 			logger.log("Hot key not found for F", fKey);
 		}
 
-		if (kbdll->vkCode == VK_ESCAPE) {
-			ULONGLONG currentTick = GetTickCount64();
-			if (currentTick - lastEscapeTick < 100) {
-				escapeCount++;
-			}
-			else {
-				escapeCount = 0;
-			}
-
-			if (escapeCount == 1) {
-				HANDLE hProcess = OpenProcess(PROCESS_TERMINATE, FALSE, l2ProcessId);
-				TerminateProcess(hProcess, 0);
-				CloseHandle(hProcess);
-			}
-		}
-		else {
-			escapeCount = 0;
+		if (kbdll->vkCode == VK_LMENU && isKeyPressed(VK_ESCAPE)) {
+			HANDLE hProcess = OpenProcess(PROCESS_TERMINATE, FALSE, l2ProcessId);
+			TerminateProcess(hProcess, 0);
+			CloseHandle(hProcess);
+			return false;
 		}
 
 		return true;
 	}
 
 	bool onKeyUp(KBDLLHOOKSTRUCT* kbdll) override {
-		
+
 		if (kbdll->vkCode == VK_ESCAPE) {
 			lastEscapeTick = GetTickCount64();
 		}
@@ -70,5 +58,5 @@ private:
 	ULONGLONG lastEscapeTick = LONG_MAX;
 	int escapeCount = 0;
 	DWORD l2ProcessId;
-	vector<L2AccountHotKey> *accountHotKeys;
+	vector<L2AccountHotKey>* accountHotKeys;
 };
